Enable Qualcomm MDP for HWC
Motorola left it disable, probably due to an error.
Using the HW compositor is always better,
since it saves power and releases GPU resources.
I measured:
- 2%+ Benchmarks
- 10%- Power usage on videos replay

I tested with a custom CM build on osprey. Working great,
but please check on other versions.
Maybe it is needed to disable the partial MDP in order to get it error free
under other models.

I checked other 8916 branches, and in every other is enabled.
https://github.com/CyanogenMod/android_device_asus_msm8916-common/blob/cm-13.0/system.prop
https://github.com/CyanogenMod/android_device_huawei_msm8916-common/blob/cm-13.0/system.prop
https://github.com/CyanogenMod/android_device_qcom_msm8916_64/blob/LA.BR.1.2.6_rb1.12/system.prop

Change-Id: I89ff31ad53c2cbe041a06433969be3176ebcf45f